+++ commands/interval.go
@@ -1,7 +1,6 @@
 package commands
 
 import (
-	"fmt"
 	"scan2epub/lang"
 	"scan2epub/service"
 
@@ -11,24 +10,15 @@ import (
 func IntervalCommand() *cli.Command {
 	l := lang.GetLocalize()
 	return &cli.Command{
-		Name:      "inter",
-		Usage:     l.Get("interval-usage"),
-		ArgsUsage: "<cron> <chap>",
-		Aliases:   []string{"i"},
-		Action:    intervalAction,
+		Name:    "inter",
+		Usage:   l.Get("interval-usage"),
+		Aliases: []string{"i"},
+		Action:  intervalAction,
 	}
 }
 
 func intervalAction(c *cli.Context) error {
-	l := lang.GetLocalize()
-	if c.NArg() < 2 {
-		return fmt.Errorf(l.Get("no-chapter-cron"))
-	}
-
-	cronStr := c.Args().Get(0)
-	chap := c.Args().Get(1)
-
-	if err := service.CronDownloadChap(cronStr, chap); err != nil {
+	if err := service.CronDownloadChap(); err != nil {
 		return err
 	}